requirements.txt
aiofiles==23.2.1
aiogram==2.25.2
aiohttp==3.8.6
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.10.0
async-timeout==4.0.3
attrs==25.4.0
Babel==2.9.1
billiard==4.2.2
celery==5.3.4
certifi==2025.10.5
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.4.4
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==3.2.16
django-cors-headers==4.3.1
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
frozenlist==1.8.0
idna==3.11
kombu==5.5.4
magic-filter==1.0.12
msgpack==1.1.2
multidict==6.7.0
packaging==25.0
prompt_toolkit==3.0.52
propcache==0.4.1
pydantic==2.8.2
pydantic_core==2.20.1
pyTelegramBotAPI==4.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
redis==5.0.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
yarl==1.22.0
